third_party/blink/web_tests/platform/generic/fast/filesystem/workers/file-entry-to-uri-sync-expected.txt

[Worker] Obtaining URL from FileEntry.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/file-entry-to-uri-sync.js
PASS [Worker] testFileEntry.toURL() is 'filesystem:file:///temporary/testFileEntry.txt'
PASS successfullyParsed is true

TEST COMPLETE
